requirements.txt
streamlit>=1.33
openai>=1.3.0
python-docx
pypdf
pandas
numpy
tqdm
